refactor(soql): address browser harness review
