+++ packages/soql-builder-ui/src/application.ts
@@ -9,7 +9,6 @@ import type { SoqlBuilderService } from './effect/soqlBuilderService.js';
 import * as Effect from 'effect/Effect';
 import * as Fiber from 'effect/Fiber';
 import * as Layer from 'effect/Layer';
-import * as ManagedRuntime from 'effect/ManagedRuntime';
 import * as Queue from 'effect/Queue';
 import * as Schema from 'effect/Schema';
 import * as Stream from 'effect/Stream';
@@ -35,8 +34,7 @@ const isSoqlBuilderActionEvent = (event: Event): event is Event & { readonly det
   event.type === SOQL_BUILDER_ACTION_EVENT && 'detail' in event && isSoqlBuilderAction(event.detail);
 
 export class SoqlBuilderApplication {
-  private connection: Fiber.RuntimeFiber<void, SoqlBuilderServiceError> | undefined;
-  private runtime: ManagedRuntime.ManagedRuntime<SoqlBuilderController, SoqlBuilderServiceError> | undefined;
+  private sessionFiber: Fiber.RuntimeFiber<void, never> | undefined;
 
   constructor(
     private readonly view: SoqlBuilderView,
@@ -44,10 +42,8 @@ export class SoqlBuilderApplication {
   ) {}
 
   public readonly connect = (): void => {
-    if (this.connection) return;
+    if (this.sessionFiber) return;
 
-    const runtime = ManagedRuntime.make(SoqlBuilderController.Default.pipe(Layer.provide(this.serviceLayer)));
-    this.runtime = runtime;
     const reportServiceError = (error: SoqlBuilderServiceError): Effect.Effect<void> =>
       Effect.sync(() => {
         this.view.viewState = {
@@ -55,56 +51,58 @@ export class SoqlBuilderApplication {
           errorMessage: error.message
         };
       });
-    const connection = runtime.runFork(
-      Effect.gen(this, function* () {
-        const controller = yield* SoqlBuilderController;
-        const actions = yield* Queue.unbounded<SoqlBuilderAction>();
-        const actionListener: EventListener = event => {
-          if (isSoqlBuilderActionEvent(event)) actions.unsafeOffer(event.detail);
-        };
+    const session = Effect.gen(this, function* () {
+      const controller = yield* SoqlBuilderController;
+      const actions = yield* Queue.unbounded<SoqlBuilderAction>();
+      const actionListener: EventListener = event => {
+        if (isSoqlBuilderActionEvent(event)) actions.unsafeOffer(event.detail);
+      };
 
-        this.view.addEventListener(SOQL_BUILDER_ACTION_EVENT, actionListener);
-        yield* Effect.addFinalizer(() =>
+      yield* Effect.acquireRelease(
+        Effect.sync(() => this.view.addEventListener(SOQL_BUILDER_ACTION_EVENT, actionListener)),
+        () =>
           Effect.sync(() => this.view.removeEventListener(SOQL_BUILDER_ACTION_EVENT, actionListener)).pipe(
             Effect.andThen(Queue.shutdown(actions))
           )
-        );
+      );
 
-        yield* Effect.all(
-          [
-            controller.states.pipe(
-              Stream.runForEach(state =>
-                Effect.sync(() => {
-                  this.view.viewState = state;
-                })
-              )
-            ),
-            Stream.fromQueue(actions).pipe(Stream.runForEach(controller.dispatch))
-          ],
-          { concurrency: 'unbounded', discard: true }
-        );
-      }).pipe(Effect.scoped, Effect.catchAll(reportServiceError))
+      yield* Effect.all(
+        [
+          controller.states.pipe(
+            Stream.runForEach(state =>
+              Effect.sync(() => {
+                this.view.viewState = state;
+              })
+            )
+          ),
+          Stream.fromQueue(actions).pipe(Stream.runForEach(controller.dispatch))
+        ],
+        { concurrency: 'unbounded', discard: true }
+      );
+    }).pipe(
+      Effect.provide(SoqlBuilderController.Default.pipe(Layer.provide(this.serviceLayer))),
+      Effect.scoped,
+      Effect.catchTags({
+        InvalidSoqlBuilderMetadataError: reportServiceError,
+        SoqlBuilderMessageChannelError: reportServiceError,
+        SoqlBuilderQueryError: reportServiceError
+      })
     );
-    this.connection = connection;
-    connection.addObserver(() => {
-      if (this.connection !== connection) return;
-      this.connection = undefined;
-      this.runtime = undefined;
-      // addObserver's callback must stay synchronous (Fiber observer contract), so runtime teardown is
-      // fire-and-forget here; a completed application fiber has no caller to receive teardown failures.
-      void runtime.dispose().catch(() => {
-        // teardown failures during disposal are non-actionable
-      });
-    });
+    const sessionFiber = Effect.yieldNow().pipe(
+      Effect.andThen(session),
+      Effect.ensuring(
+        Effect.sync(() => {
+          if (this.sessionFiber === sessionFiber) this.sessionFiber = undefined;
+        })
+      ),
+      Effect.runFork
+    );
+    this.sessionFiber = sessionFiber;
   };
 
   public readonly disconnect = async (): Promise<void> => {
-    const connection = this.connection;
-    const runtime = this.runtime;
-    this.connection = undefined;
-    this.runtime = undefined;
-    if (!runtime) return;
-    if (connection) await runtime.runPromise(Fiber.interrupt(connection));
-    await runtime.dispose();
+    const sessionFiber = this.sessionFiber;
+    this.sessionFiber = undefined;
+    if (sessionFiber) await sessionFiber.pipe(Fiber.interrupt, Effect.runPromise);
   };
 }
