+++ packages/soql-builder-ui/src/testing/fakeSoqlBuilderService.ts
@@ -5,20 +5,17 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import type { SoqlBuilderAction, SoqlBuilderServiceError, SoqlBuilderState } from '../domain.js';
+import type { SoqlBuilderState } from '../domain.js';
 import type * as Duration from 'effect/Duration';
+import type * as Effect from 'effect/Effect';
 import { SoqlBuilderService } from '../effect/soqlBuilderService.js';
-import { makeFakeEffectDriver, type FakeEffectDriver, type FakeEffectDriverStats } from './fakeEffectDriver.js';
+import { makeFakeEffectService, type FakeEffectServiceStats } from './fakeEffectService.js';
 
-export type FakeSoqlBuilderServiceStats = FakeEffectDriverStats;
-export type FakeSoqlBuilderService = FakeEffectDriver<
-  SoqlBuilderService,
-  SoqlBuilderState,
-  SoqlBuilderAction,
-  SoqlBuilderServiceError
->;
+export type FakeSoqlBuilderServiceStats = FakeEffectServiceStats;
 
 export const makeFakeSoqlBuilderService = (
   initialState: SoqlBuilderState,
   options: { readonly dispatchLatency?: Duration.DurationInput } = {}
-) => makeFakeEffectDriver(SoqlBuilderService, initialState, options);
+) => makeFakeEffectService(SoqlBuilderService, initialState, options);
+
+export type FakeSoqlBuilderService = Effect.Effect.Success<ReturnType<typeof makeFakeSoqlBuilderService>>;
